refactored apps (extracted AbstractApp from the four app classes, which they now all extend)
