+++ sizes.txt
@@ -1,5 +1,5 @@
-uSize:2:
-aSize:1:
-eSize:1:
+uSize:0:
+aSize:0:
+eSize:0:
 sSize:0:
 bSize:0:
