+++ src/interfaces/AppInterface.java
@@ -5,6 +5,4 @@ import java.io.Serializable;
 public interface AppInterface extends Serializable {
   
   void displayMainMenu();
-
-//  void lodgeAComplaint();
 }
